requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
numpy==1.23.3
Pillow==9.2.0
pydantic==1.10.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tflite-runtime==2.10.0
typing_extensions==4.3.0
uvicorn==0.18.3
